Move selectors to views manager
