+++ src/kubernetes/kinds/kind.rs
@@ -4,6 +4,7 @@ use crate::{
 };
 
 /// Represents kubernetes kind.
+#[derive(Clone)]
 pub struct KindItem {
     pub uid: Option<String>,
     pub group: String,
